Rename prime fields directory to prime_base and prime_extension
