+++ libff/algebra/curves/bn128/bn128_init.hpp
@@ -10,7 +10,7 @@
 #include "depends/ate-pairing/include/bn.h"
 
 #include <libff/algebra/curves/public_params.hpp>
-#include <libff/algebra/fields/prime/fp.hpp>
+#include <libff/algebra/fields/prime_base/fp.hpp>
 
 namespace libff {
