+++ libff/algebra/curves/mnt/mnt6/mnt6_init.hpp
@@ -14,9 +14,9 @@
 
 #include <libff/algebra/curves/mnt/mnt46_common.hpp>
 #include <libff/algebra/curves/public_params.hpp>
-#include <libff/algebra/fields/prime/fp.hpp>
-#include <libff/algebra/fields/prime/fp3.hpp>
-#include <libff/algebra/fields/prime/fp6_2over3.hpp>
+#include <libff/algebra/fields/prime_base/fp.hpp>
+#include <libff/algebra/fields/prime_extension/fp3.hpp>
+#include <libff/algebra/fields/prime_extension/fp6_2over3.hpp>
 
 namespace libff {
