+++ libff/algebra/fields/prime_extension/fp3.tcc
@@ -10,7 +10,7 @@
 #ifndef FP3_TCC_
 #define FP3_TCC_
 
-#include <libff/algebra/fields/prime/field_utils.hpp>
+#include <libff/algebra/fields/prime_base/field_utils.hpp>
 
 namespace libff {
