+++ libff/algebra/scalar_multiplication/multiexp.tcc
@@ -19,7 +19,7 @@
 #include <type_traits>
 
 #include <libff/algebra/fields/bigint.hpp>
-#include <libff/algebra/fields/prime/fp_aux.tcc>
+#include <libff/algebra/fields/prime_base/fp_aux.tcc>
 #include <libff/algebra/scalar_multiplication/multiexp.hpp>
 #include <libff/algebra/scalar_multiplication/wnaf.hpp>
 #include <libff/common/profiling.hpp>
